Fix a panic when Insert/Update is called with ptr struct
The Insert/Update functions were missing an indirectV (which resolves
any pointer indirections away from a reflect.Value). This caused
value.FieldByName to be called, which panics on a non-struct type.

The error was not caught successfully because sqlToGoFields, which
should catch the type error, indirects the type.

This fix makes Insert/Update also indirect the value so they work with
both pointers and by-value structs. It also adds a testcase for the
issue.