Add support for time.Time fields marshalled to timestamps.
A lot of legacy databases don't use built-in SQL time types to store
timestamp data; instead they opt to store UNIX timestamps in numeric
fields. This patch adds functionality to tag Time fields in Go structs
with "unix", which enables marshalling to/from timestamps when talking
to the SQL database.

The marshalling is performed in Go to avoid RDBMS-specific
functionality.